CSS DSL: fix applying rules from Promise to DomSideEffect
