+++ compose/web/src/jsMain/kotlin/androidx/compose/web/css/CSSRules.kt
@@ -16,14 +16,7 @@
 
 package androidx.compose.web.css
 
-import androidx.compose.web.attributes.AttrsBuilder
-import androidx.compose.web.attributes.Tag
 import androidx.compose.web.css.selectors.CSSSelector
-import org.w3c.dom.HTMLStyleElement
-import org.w3c.dom.css.CSSGroupingRule
-import org.w3c.dom.css.CSSRule
-import org.w3c.dom.css.StyleSheet
-import kotlin.js.Promise
 
 interface CSSStyleRuleBuilder : StyleBuilder
 
@@ -59,73 +52,6 @@ abstract class CSSGroupingRuleDeclaration(
 
 typealias CSSRuleDeclarationList = List<CSSRuleDeclaration>
 typealias MutableCSSRuleDeclarationList = MutableList<CSSRuleDeclaration>
-
-private fun StyleSheet.addRule(cssRule: String): CSSRule {
-    val cssRuleIndex = this.insertRule(cssRule, this.cssRules.length)
-    return this.cssRules[cssRuleIndex]
-}
-
-private fun CSSGroupingRule.addRule(cssRule: String): CSSRule {
-    val cssRuleIndex = this.insertRule(cssRule, this.cssRules.length)
-    return this.cssRules[cssRuleIndex]
-}
-
-private fun StyleSheet.addRule(cssRuleDeclaration: CSSRuleDeclaration) {
-    val cssRule = addRule("${cssRuleDeclaration.header} {}")
-    fillRule(cssRuleDeclaration, cssRule)
-}
-
-private fun CSSGroupingRule.addRule(cssRuleDeclaration: CSSRuleDeclaration) {
-    val cssRule = addRule("${cssRuleDeclaration.header} {}")
-    fillRule(cssRuleDeclaration, cssRule)
-}
-
-private fun fillRule(
-    cssRuleDeclaration: CSSRuleDeclaration,
-    cssRule: CSSRule
-) {
-    when (cssRuleDeclaration) {
-        is CSSStyleRuleDeclaration ->
-            cssRuleDeclaration.properties.forEach { (name, value) ->
-                cssRule.styleMap.set(name, value)
-            }
-        is CSSGroupingRuleDeclaration -> {
-            val cssGroupingRule = cssRule.unsafeCast<CSSGroupingRule>()
-            cssRuleDeclaration.rules.forEach { childRuleDeclaration ->
-                cssGroupingRule.addRule(childRuleDeclaration)
-            }
-        }
-    }
-}
-
-fun AttrsBuilder<Tag.Style>.cssRules(rules: CSSRuleDeclarationList): AttrsBuilder<Tag.Style> {
-    prop(setCSSRules, rules)
-    return this
-}
-
-private val setCSSRules: (HTMLStyleElement, CSSRuleDeclarationList) -> Unit = { e, cssRules ->
-    // TODO: Make CSSRule as composable child of Style element instead of replacing each time
-    val sheet = e.sheet
-    if (sheet == null) {
-        // schedule after element is attached
-        Promise.resolve(Unit).then {
-            e.sheet?.let {
-                setCSSRules(it, cssRules)
-            }
-        }
-    } else {
-        setCSSRules(sheet, cssRules)
-    }
-}
-
-private fun setCSSRules(sheet: StyleSheet, cssRules: CSSRuleDeclarationList) {
-    repeat(sheet.cssRules.length) {
-        sheet.deleteRule(0)
-    }
-    cssRules.forEach { cssRule ->
-        sheet.addRule(cssRule)
-    }
-}
 
 fun buildCSSStyleRule(cssRule: CSSStyleRuleBuilder.() -> Unit): StylePropertyList {
     val builder = CSSRuleBuilderImpl()
