+++ compose/web/src/jsMain/kotlin/androidx/compose/web/elements/Style.kt
@@ -19,11 +19,21 @@ package androidx.compose.web.elements
 import androidx.compose.runtime.Composable
 import androidx.compose.web.attributes.AttrsBuilder
 import androidx.compose.web.attributes.Tag
+import androidx.compose.web.css.CSSGroupingRuleDeclaration
+import androidx.compose.web.css.CSSRuleDeclaration
 import androidx.compose.web.css.CSSRuleDeclarationList
+import androidx.compose.web.css.CSSStyleRuleDeclaration
 import androidx.compose.web.css.StyleSheetBuilder
 import androidx.compose.web.css.StyleSheetBuilderImpl
 import androidx.compose.web.css.cssRules
+import androidx.compose.web.css.deleteRule
+import androidx.compose.web.css.get
+import androidx.compose.web.css.insertRule
+import androidx.compose.web.css.styleMap
 import org.w3c.dom.HTMLStyleElement
+import org.w3c.dom.css.CSSGroupingRule
+import org.w3c.dom.css.CSSRule
+import org.w3c.dom.css.StyleSheet
 
 @Composable
 inline fun Style(
@@ -43,10 +53,67 @@ inline fun Style(
     TagElement<Tag.Style, HTMLStyleElement>(
         tagName = "style",
         applyAttrs = {
-            cssRules(cssRules)
             applyAttrs()
         },
-        applyStyle = {},
-        content = {}
-    )
+        applyStyle = {}
+    ) {
+        DomSideEffect(cssRules) { style ->
+            style.sheet?.let { sheet ->
+                setCSSRules(sheet, cssRules)
+                onDispose {
+                    clearCSSRules(sheet)
+                }
+            }
+        }
+    }
+}
+
+fun clearCSSRules(sheet: StyleSheet) {
+    repeat(sheet.cssRules.length) {
+        sheet.deleteRule(0)
+    }
+}
+
+fun setCSSRules(sheet: StyleSheet, cssRules: CSSRuleDeclarationList) {
+    cssRules.forEach { cssRule ->
+        sheet.addRule(cssRule)
+    }
+}
+
+private fun StyleSheet.addRule(cssRule: String): CSSRule {
+    val cssRuleIndex = this.insertRule(cssRule, this.cssRules.length)
+    return this.cssRules[cssRuleIndex]
+}
+
+private fun CSSGroupingRule.addRule(cssRule: String): CSSRule {
+    val cssRuleIndex = this.insertRule(cssRule, this.cssRules.length)
+    return this.cssRules[cssRuleIndex]
+}
+
+private fun StyleSheet.addRule(cssRuleDeclaration: CSSRuleDeclaration) {
+    val cssRule = addRule("${cssRuleDeclaration.header} {}")
+    fillRule(cssRuleDeclaration, cssRule)
+}
+
+private fun CSSGroupingRule.addRule(cssRuleDeclaration: CSSRuleDeclaration) {
+    val cssRule = addRule("${cssRuleDeclaration.header} {}")
+    fillRule(cssRuleDeclaration, cssRule)
+}
+
+private fun fillRule(
+    cssRuleDeclaration: CSSRuleDeclaration,
+    cssRule: CSSRule
+) {
+    when (cssRuleDeclaration) {
+        is CSSStyleRuleDeclaration ->
+            cssRuleDeclaration.properties.forEach { (name, value) ->
+                cssRule.styleMap.set(name, value)
+            }
+        is CSSGroupingRuleDeclaration -> {
+            val cssGroupingRule = cssRule.unsafeCast<CSSGroupingRule>()
+            cssRuleDeclaration.rules.forEach { childRuleDeclaration ->
+                cssGroupingRule.addRule(childRuleDeclaration)
+            }
+        }
+    }
 }
